+++ modules/apps/wiki-engine-mediawiki/wiki-engine-mediawiki/src/main/java/com/liferay/wiki/engine/mediawiki/internal/LiferaySearchEngine.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.engine.mediawiki;
+package com.liferay.wiki.engine.mediawiki.internal;
 
 import java.util.List;
