+++ modules/apps/wiki-engine-mediawiki/wiki-engine-mediawiki/src/main/java/com/liferay/wiki/engine/mediawiki/internal/matchers/EditURLMatcher.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.engine.mediawiki.matchers;
+package com.liferay.wiki.engine.mediawiki.internal.matchers;
 
 import javax.portlet.PortletURL;
