+++ modules/apps/wiki-engine-mediawiki/wiki-engine-mediawiki/src/main/resources/jamwiki.properties
@@ -1,3 +1,3 @@
-database-type=com.liferay.wiki.engine.mediawiki.LiferayDataHandler
+database-type=com.liferay.wiki.engine.mediawiki.internal.LiferayDataHandler
 props-initialized=true
-search-engine=com.liferay.wiki.engine.mediawiki.LiferaySearchEngine
+search-engine=com.liferay.wiki.engine.mediawiki.internal.LiferaySearchEngine
